requirements.txt
boto3>=1.28.0
botocore>=1.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
